+++ gc-testimonials-to-testimonials.php
@@ -1,9 +1,9 @@
 <?php
 /**
- * Plugin Name: GC Testimonials to Testimonials by Aihrus
+ * Plugin Name: Testimonials – GC Testimonials Migrator
  * Plugin URI: http://wordpress.org/plugins/gc-testimonials-to-testimonials/
  * Description: Migrate GC Testimonials entries to Testimonials by Aihrus custom post types.
- * Version: 1.2.0
+ * Version: 1.2.1
  * Author: Michael Cannon
  * Author URI: http://aihr.us/resume/
  * License: GPLv2 or later
@@ -35,8 +35,8 @@ define( 'GCT2T_REQ_BASE', 'testimonials-widget/testimonials-widget.php' );
 define( 'GCT2T_REQ_BASE_PREM', 'testimonials-widget-premium/testimonials-widget-premium.php' );
 define( 'GCT2T_REQ_NAME', 'Testimonials by Aihrus' );
 define( 'GCT2T_REQ_SLUG', 'testimonials-widget' );
-define( 'GCT2T_REQ_VERSION', '2.18.0' );
-define( 'GCT2T_VERSION', '1.2.0' );
+define( 'GCT2T_REQ_VERSION', '2.18.1' );
+define( 'GCT2T_VERSION', '1.2.1' );
 
 if ( defined( 'TW_DIR_LIB' ) ) {
 	define( 'GCT2T_DIR_LIB_ALT', TW_DIR_LIB );